refactor: refactor routes and routes handle names
